+++ libethereum/ChainParams.cpp
@@ -115,7 +115,7 @@ ChainParams ChainParams::loadConfig(
     setOptionalU256Parameter(cp.durationLimit, c_durationLimit);
 
     if (params.count(c_chainID))
-        cp.chainID = int(fromBigEndian<u256>(fromHex(params.at(c_chainID).get_str())));
+        cp.chainID = uint64_t(fromBigEndian<u256>(fromHex(params.at(c_chainID).get_str())));
     if (params.count(c_networkID))
         cp.networkID = int(fromBigEndian<u256>(fromHex(params.at(c_networkID).get_str())));
     cp.allowFutureBlocks = params.count(c_allowFutureBlocks);
